Revert "Display badge if user is in subscription"
