--- .nvmrc
@@ -1,1 +1,0 @@
-10.15.0
